Made the V26 projects work without nuget, and renamed/renamespaced for clarity
- Renamed V26/StoreAndForwardNETClient to StoreAndForwardClient
- Renamed V26/ViewsNETClient to ViewsClient
- Namespaced V26/StoreAndForwardClient and V26/ViewsClient for consistency
- fixed namespace in calc plugin
- Adjusted readme
