+++ V26/StoreAndForwardClient/StoreAndForwardClient.csproj
@@ -1,0 +1,36 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <RootNamespace>$(MSBuildProjectName)</RootNamespace>
+    <AssemblyName>$(MSBuildProjectName)</AssemblyName>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Reference Include="Google.Protobuf">
+      <HintPath>..\..\DLLs\Google.Protobuf.dll</HintPath>
+    </Reference>
+    <Reference Include="Grpc.Core.Api">
+      <HintPath>..\..\DLLs\Grpc.Core.Api.dll</HintPath>
+    </Reference>
+    <Reference Include="System.Text.Json">
+      <HintPath>..\..\DLLs\System.Text.Json.dll</HintPath>
+    </Reference>
+    <Reference Include="Canary.StoreAndForward2.Grpc.Api">
+      <HintPath>..\..\DLLs\Canary.StoreAndForward2.Grpc.Api.dll</HintPath>
+    </Reference>
+    <Reference Include="Canary.StoreAndForward2.Grpc.Api.Helper">
+      <HintPath>..\..\DLLs\Canary.StoreAndForward2.Grpc.Api.Helper.dll</HintPath>
+    </Reference>
+    <Reference Include="Canary.Utility.LogCore">
+      <HintPath>..\..\DLLs\Canary.Utility.LogCore.dll</HintPath>
+    </Reference>
+    <Reference Include="Canary.Utility.ProtobufSharedTypes">
+      <HintPath>..\..\DLLs\Canary.Utility.ProtobufSharedTypes.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
+</Project>
