--- V26/StoreAndForwardNETClient/StoreAndForwardNETClient.csproj
@@ -1,61 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Google.Protobuf" Version="3.33.4" />
-    <PackageReference Include="Grpc.Core.Api" Version="2.76.0" />
-    <PackageReference Include="Grpc.Net.Client" Version="2.76.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.4" />
-    <PackageReference Include="System.Text.Json" Version="10.0.2" />
-    <PackageReference Include="System.Threading.AccessControl" Version="10.0.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="Canary.DiagnosticHealthSystem.Shared">
-      <HintPath>..\..\DLLs\Canary.DiagnosticHealthSystem.Shared.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.StoreAndForward2.Grpc.Api">
-      <HintPath>..\..\DLLs\Canary.StoreAndForward2.Grpc.Api.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.StoreAndForward2.Grpc.Api.Helper">
-      <HintPath>..\..\DLLs\Canary.StoreAndForward2.Grpc.Api.Helper.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.CanaryClientHelper">
-      <HintPath>..\..\DLLs\Canary.Utility.CanaryClientHelper.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.CanaryShared">
-      <HintPath>..\..\DLLs\Canary.Utility.CanaryShared.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.GrpcHelper">
-      <HintPath>..\..\DLLs\Canary.Utility.GrpcHelper.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.JsonCore">
-      <HintPath>..\..\DLLs\Canary.Utility.JsonCore.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.LogCore">
-      <HintPath>..\..\DLLs\Canary.Utility.LogCore.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.PathCore">
-      <HintPath>..\..\DLLs\Canary.Utility.PathCore.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.ProtobufSharedTypes">
-      <HintPath>..\..\DLLs\Canary.Utility.ProtobufSharedTypes.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.SecurityCore">
-      <HintPath>..\..\DLLs\Canary.Utility.SecurityCore.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.SettingsCore">
-      <HintPath>..\..\DLLs\Canary.Utility.SettingsCore.dll</HintPath>
-    </Reference>
-    <Reference Include="Canary.Utility.TaskCore">
-      <HintPath>..\..\DLLs\Canary.Utility.TaskCore.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-</Project>
